fix/refactory: fix validation error on updating entities + modularization of Request classes
the rules for validating data send to update a bill, for example, should be different than the ones from creating: on updating the validation rules should be optional. Also the Request classes have been modularized (e.g, Request\Bill\BillStoreRequest).
