+++ app/Http/Requests/Bill/BillUpdateRequest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace App\Http\Requests\Bill;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class BillUpdateRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     */
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
+     */
+    public function rules(): array
+    {
+        return [
+            'title' => 'string|max:255',
+            'description' => 'nullable|string|max:65535',
+            'amount' => 'numeric',
+            'due_date' => 'date',
+        ];
+    }
+}
