+++ app/Http/Requests/Task/TaskUpdateRequest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace App\Http\Requests\Task;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class TaskUpdateRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     */
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
+     */
+    public function rules(): array
+    {
+        return [
+            'task_category_id' => 'nullable|exists:task_categories,id',
+            'title' => 'string|max:255',
+            'description' => 'nullable|string|max:65535',
+            'due_date' => 'date',
+        ];
+    }
+}
