+++ app/Http/Requests/Transaction/TransactionStoreRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Transaction;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class BillRequest extends FormRequest
+class TransactionStoreRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -22,10 +22,13 @@ class BillRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'title' => 'required|string|max:255',
-            'description' => 'nullable|string|max:65535',
-            'amount' => 'numeric',
-            'due_date' => 'required|date',
+            'user_id' => 'exists:users,id',
+            'bill_id' => 'exists:bills,id',
+            'transaction_category_id' =>
+                'nullable|exists:transaction_categories,id',
+            'amount' => 'required|numeric',
+            'type' => 'string|in:income,expense',
+            'description' => 'required|string|max:65535',
         ];
     }
 }
